HASH SET IN HACKER RANK
According to the question we have to take a pair of names and see if they are the same. 
The count remains the same if they are the same.
But if a different pair comes the count changes. 
Each time we check a pair the count number must be printed that is the size of the hashset. 
 So my solution was to just simply join the words and then compare them and print the hash set size later again!!

glad it worked 😱😭😂!!